+++ app/src/main/java/example/com/myapplication/dto/Version.kt
@@ -1,0 +1,10 @@
+package example.com.myapplication.dto
+
+import android.os.Parcelable
+import com.google.gson.annotations.SerializedName
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+data class Version(
+        @SerializedName("msg") val msg: String
+) : Parcelable
